First attempt at differentiating between legacy and modern editor UIs.
