+++ plugins/cad/src/components/ResourceEditorDialog/ModernResourceEditorDialog.tsx
@@ -1,0 +1,27 @@
+/**
+ * Copyright 2024 The Nephio Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { Dialog } from '@material-ui/core';
+import React from 'react';
+import { ResourceEditorDialogProps } from './ResourceEditorDialog';
+
+export const ModernResourceEditorDialog = ({ open, onClose }: ResourceEditorDialogProps) => {
+  const onDialogClose = (): void => {
+    onClose();
+  };
+
+  return <Dialog open={open} onClose={onDialogClose} maxWidth="lg" />;
+};
