+++ plugins/cad/src/components/ResourceEditorDialog/ResourceEditorDialog.tsx
@@ -14,18 +14,17 @@
  * limitations under the License.
  */
 
-import { Button, Dialog, DialogActions, DialogContent, DialogTitle, makeStyles } from '@material-ui/core';
-import React, { useCallback, useEffect, useState } from 'react';
+import React from 'react';
 import { KubernetesResource } from '../../types/KubernetesResource';
 import { getGroupVersionKind } from '../../utils/kubernetesResource';
 import { PackageResource } from '../../utils/packageRevisionResources';
 import { loadYaml } from '../../utils/yaml';
-import { YamlViewer } from '../Controls';
-import { FirstClassEditorSelector } from './components/FirstClassEditorSelector';
+import { LegacyResourceEditorDialog } from './LegacyResourceEditorDialog';
+import { ModernResourceEditorDialog } from './ModernResourceEditorDialog';
 
 type OnSaveYamlFn = (yaml: string) => void;
 
-type ResourceEditorProps = {
+export type ResourceEditorDialogProps = {
   open: boolean;
   onClose: () => void;
   yaml: string;
@@ -33,93 +32,20 @@ type ResourceEditorProps = {
   packageResources: PackageResource[];
 };
 
-const useStyles = makeStyles({
-  container: {
-    width: '660px',
-    minHeight: 'min(420px, calc(80vh - 120px))',
-    maxHeight: 'calc(80vh - 120px)',
-    marginBottom: '10px',
-    overflowY: 'scroll',
-  },
-});
+const MODERN_EDITOR_RESOURCES = [
+  'infra.nephio.org/v1alpha1/Token',
+  'infra.nephio.org/v1alpha1/Network',
+  'infra.nephio.org/v1alpha1/WorkloadCluster',
+  'req.nephio.org/v1alpha1/Capacity',
+];
 
-export const ResourceEditorDialog = ({ open, onClose, yaml, onSaveYaml, packageResources }: ResourceEditorProps) => {
-  const classes = useStyles();
-
-  const [showYamlView, setShowYamlView] = useState<boolean>(false);
-  const [latestYaml, setLatestYaml] = useState<string>('');
-  const [isNamedEditor, setIsNamedEditor] = useState<boolean>(true);
-
-  const resourceYaml = loadYaml(yaml) as KubernetesResource;
+export const ResourceEditorDialog: React.FC<ResourceEditorDialogProps> = props => {
+  const resourceYaml = loadYaml(props.yaml) as KubernetesResource;
   const groupVersionKind = resourceYaml && getGroupVersionKind(resourceYaml);
 
-  useEffect(() => {
-    if (open) {
-      // Reset the state of the dialog anytime it is opened
-      setLatestYaml(yaml);
-      setShowYamlView(false);
-      setIsNamedEditor(true);
-    }
-  }, [open, yaml]);
-
-  const toggleView = (): void => {
-    setShowYamlView(!showYamlView);
-  };
-
-  const handleNoNamedEditor = useCallback(() => setIsNamedEditor(false), []);
-
-  const handleUpdatedYaml = (updatedYaml: string): void => {
-    setLatestYaml(updatedYaml);
-  };
-
-  const onSave = (): void => {
-    onSaveYaml(latestYaml);
-    onClose();
-  };
-
-  const onDialogClose = (): void => {
-    onClose();
-  };
-
-  const latestYamlHeight = (latestYaml.split('\n').length + 1) * 18;
-  const title =
-    resourceYaml?.kind && resourceYaml?.metadata
-      ? `${resourceYaml.kind} ${resourceYaml?.metadata.name}`
-      : 'New Resource';
-
-  return (
-    <Dialog open={open} onClose={onDialogClose} maxWidth="lg">
-      <DialogTitle>{title}</DialogTitle>
-      <DialogContent>
-        <>
-          <div className={classes.container} style={{ height: `${latestYamlHeight}px` }}>
-            {!showYamlView && isNamedEditor ? (
-              <FirstClassEditorSelector
-                groupVersionKind={groupVersionKind}
-                yaml={latestYaml}
-                packageResources={packageResources}
-                onNoNamedEditor={handleNoNamedEditor}
-                onUpdatedYaml={handleUpdatedYaml}
-              />
-            ) : (
-              <YamlViewer value={latestYaml} allowEdit onUpdatedValue={handleUpdatedYaml} />
-            )}
-          </div>
-          {isNamedEditor && (
-            <Button variant="text" color="primary" onClick={toggleView}>
-              Show {showYamlView ? 'Formatted View' : 'YAML View'}
-            </Button>
-          )}
-        </>
-      </DialogContent>
-      <DialogActions>
-        <Button color="primary" onClick={onDialogClose}>
-          Cancel
-        </Button>
-        <Button color="primary" onClick={onSave}>
-          Save
-        </Button>
-      </DialogActions>
-    </Dialog>
+  return MODERN_EDITOR_RESOURCES.includes(groupVersionKind) ? (
+    <ModernResourceEditorDialog {...props} />
+  ) : (
+    <LegacyResourceEditorDialog {...props} />
   );
 };
